Move the multi-GPU launcher into isaaclab_rl
The launcher lived entirely in scripts/reinforcement_learning, which left
it outside the package that already owns the train and play entry points
and made its logic unavailable to anything but that one script.

Move it to isaaclab_rl.entrypoints.multigpu, exposed as
run_train_multigpu_cli, and reduce the script to a shim like train.py and
play.py. The module doubles as the per-rank trainer: the launcher points
torchrun at this file, so its __main__ block is what each rank runs. That
keeps the worker resolvable from the package rather than from a path
relative to the script, which matters for wheel installs.

The list of distributed-capable backends moves to api.MULTI_GPU_BACKENDS
so it sits next to BackendName instead of being restated in the launcher,
and the skrl JAX exclusion list is now derived from the torchrun options
rather than spelled out a second time.
