+++ source/isaaclab/test/cli/test_train_multigpu_command_building.py
@@ -25,7 +25,6 @@ without a GPU or Isaac Sim installation.
 from __future__ import annotations
 
 import argparse
-import importlib.util
 import shlex
 import sys
 from pathlib import Path
@@ -34,14 +33,7 @@ import pytest
 
 from isaaclab.app.app_launcher import AppLauncher
 
-# The launcher script lives outside the installed packages; load it by path.
-# This test lives at source/isaaclab/test/cli/test_train_multigpu_command_building.py.
-_REPO_ROOT = Path(__file__).resolve().parents[4]
-_TRAIN_MULTIGPU_PATH = _REPO_ROOT / "scripts" / "reinforcement_learning" / "train_multigpu.py"
-
-_spec = importlib.util.spec_from_file_location("train_multigpu", _TRAIN_MULTIGPU_PATH)
-train_multigpu = importlib.util.module_from_spec(_spec)
-_spec.loader.exec_module(train_multigpu)
+from isaaclab_rl.entrypoints import multigpu as train_multigpu
 
 
 def _build_command(argv: list[str]) -> list[str]:
@@ -52,7 +44,7 @@ def _build_command(argv: list[str]) -> list[str]:
 
 def _forwarded_train_argv(command: list[str]) -> list[str]:
     """Return the argv forwarded to the child training script."""
-    return command[command.index(str(train_multigpu.TRAIN_SCRIPT)) + 1 :]
+    return command[command.index(train_multigpu.WORKER_SCRIPT) + 1 :]
 
 
 def _parse_as_training_script(child_argv: list[str], monkeypatch: pytest.MonkeyPatch) -> str:
@@ -203,7 +195,9 @@ class TestKitArgsForwarding:
         assert kit_args == "--foo=/bar"
 
     def test_dry_run_prints_shell_parsable_command(self, capsys):
-        exit_code = train_multigpu.main(["--dry_run", "--task", "Isaac-Cartpole-Direct", "--kit_args", "--foo=/bar"])
+        exit_code = train_multigpu.run_train_multigpu_cli(
+            ["--dry_run", "--task", "Isaac-Cartpole-Direct", "--kit_args", "--foo=/bar"]
+        )
         assert exit_code == 0
         printed = capsys.readouterr().out.strip()
         tokens = shlex.split(printed)
