+++ source/isaaclab_rl/changelog.d/multigpu-entrypoint-refactor.minor.rst
@@ -1,0 +1,6 @@
+Added
+^^^^^
+
+* Added :func:`~isaaclab_rl.entrypoints.run_train_multigpu_cli`, which moves the multi-GPU launcher
+  into the package alongside the train and play entry points.
+  ``scripts/reinforcement_learning/train_multigpu.py`` is now a shim over it.
