+++ source/isaaclab_rl/isaaclab_rl/entrypoints/__init__.py
@@ -25,6 +25,7 @@ Example:
 
 from .api import BackendName, PlaybackRequest, SimpleAgentRequest, TrainingRequest, play, random_agent, train, zero_agent
 from .dispatch import run_play_cli, run_random_agent_cli, run_train_cli, run_zero_agent_cli
+from .multigpu import run_train_multigpu_cli
 
 __all__ = [
     "BackendName",
@@ -36,6 +37,7 @@ __all__ = [
     "run_play_cli",
     "run_random_agent_cli",
     "run_train_cli",
+    "run_train_multigpu_cli",
     "run_zero_agent_cli",
     "train",
     "zero_agent",
